requirements.txt
streamlit
random
base64
panda
requests
openai